feat: refactor agent workflow and enhance market data handling
- Updated the project structure to fix  langgraph reserved src issue.
- Refactored the agent's workflow in `graph.py`, adding new nodes for market analysis, risk assessment, strategy generation, and performance analysis.
- Implemented asynchronous methods for fetching market and performance data, enhancing overall performance.
- Introduced a new `MarketDataJob` for scheduled market data fetching and saving.
- Enhanced error handling and logging across various components to improve traceability and user feedback.
- Updated `langgraph.json` to reflect changes in the workflow and entry points, ensuring consistency with the new structure.

These changes significantly improve the agent's functionality and maintainability.
