+++ app/conditions/strategy_condition.py
