+++ app/config/strategy_config.py
