+++ app/evm/contracts/abis/strategy_abi.py
