+++ app/nodes/performance_node.py
@@ -1,9 +1,9 @@
 import logging
-from src.config.llm_config import CREATIVE_TEMPERATURE, DEFAULT_TEMPERATURE, GPT3_5_MODEL
+from app.config.llm_config import CREATIVE_TEMPERATURE, DEFAULT_TEMPERATURE, GPT3_5_MODEL
 from langchain_openai import ChatOpenAI
 from langchain_core.messages import SystemMessage, HumanMessage, AIMessage
-from src.prompts.performance_prompts import PerformanceAnalysisPrompt
-from src.services.performance_service import PerformanceService
+from app.prompts.performance_prompts import PerformanceAnalysisPrompt
+from app.services.performance_service import PerformanceService
 
 logger = logging.getLogger("performance_node")
