+++ app/prompts/strategy_prompts.py
