+++ app/services/vector_service.py
