+++ app/types/strategy.py
